fix(views): enforce float typecasting on all number_format calls across billing, customer profile, and noc hardware views

+++ resources/views/finance/billing-registrasi.blade.php
@@ -281,7 +281,7 @@
                         <!-- 3. Amount -->
                         <td class="py-3.5 px-4 align-top">
                             <div class="flex items-center gap-1.5 font-bold text-white text-sm">
-                                <span>Rp {{ number_format($r->total_reg ?? $r->biaya_reg ?? 0, 0, ',', '.') }}</span>
+                                <span>Rp {{ number_format((float) ($r->total_reg ?? ($r->biaya_reg ?? 0)), 0, ',', '.') }}</span>
                                 <svg class="w-3.5 h-3.5 text-slate-500" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor">
                                     <path stroke-linecap="round" stroke-linejoin="round" d="M16.5 10.5V6.75a4.5 4.5 0 1 0-9 0v3.75m-.75 11.25h10.5a2.25 2.25 0 0 0 2.25-2.25v-6.75a2.25 2.25 0 0 0-2.25-2.25H6.75a2.25 2.25 0 0 0-2.25 2.25v6.75a2.25 2.25 0 0 0 2.25 2.25Z" />
                                 </svg>
